requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-jalali==7.4.0
django-jalali-date==2.0.0
docopt==0.6.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
num2words==0.5.14
persian-tools==0.0.11
pillow==11.3.0
requests==2.32.4
setuptools==80.9.0
smsir-python==1.0.8
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
